+++ CHANGES.rst
@@ -4,6 +4,9 @@ Release 7.3.5 (in development)
 Bugs fixed
 ----------
 
+* Re-export various objects from ``sphinx.domains.python._object``
+  in ``sphinx.domains.python``.
+  Patch by Jacob Chesslo and Adam Turner.
 
 Release 7.3.4 (released Apr 17, 2024)
 =====================================